use the same snapping strategy for both multiselect and single-select resize
